Add bloc to article card
