+++ lib/widgets/article_card/bloc/article_state.dart
@@ -1,0 +1,10 @@
+import 'package:equatable/equatable.dart';
+
+class ArticleFavoriteState extends Equatable {
+  final bool isFavorite;
+
+  const ArticleFavoriteState(this.isFavorite);
+
+  @override
+  List<Object> get props => [isFavorite];
+}
